src/my_second_homework/secondhomework.py
# tic_tac_toe.py
import pyfiglet
from unittest.mock import patch

# Create a class for the Tic-Tac-Toe game
class myClass:
    def __init__(self):
        """
        Initializes a new instance of the Tic Tac Toe game.

        The game board is initialized as a 3x3 grid with empty spaces.
        The current player is set to 'X' by default.
        The winner and draw state are initialized to None and False, respectively.
        """
        self.board = [[' ' for _ in range(3)] for _ in range(3)]
        self.current_player = 'X'
        self.winner = None
        self.draw = False

    def make_move(self, row, col):
        """
        Places the current player's symbol on the board at the specified row and column.
        If the move is valid, the current player is switched to the other player's turn.
        If the current move results in a win, the winner is set to the current player.
        If the game has ended in a draw, the draw flag is set to True.

        Args:
            row (int): The row to place the symbol in.
            col (int): The column to place the symbol in.

        Returns:
            None
        """
    def make_move(self, row, col):
        """
        Places the current player's symbol on the board at the specified row and column.
        If the move is valid, the current player is switched to the other player's turn.
        If the current move results in a win, the winner is set to the current player.
        If the game has ended in a draw, the draw flag is set to True.

        Args:
            row (int): The row to place the symbol in.
            col (int): The column to place the symbol in.

        Returns:
            None
        """
    class TicTacToe:
        """
        A class representing a Tic Tac Toe game.

        Attributes:
            board (list): A 3x3 list representing the game board.
            current_player (str): The symbol of the player who is currently taking their turn.
            winner (str): The symbol of the player who has won the game, or None if the game has not been won yet.
            draw (bool): True if the game has ended in a draw, False otherwise.
        """

        def __init__(self):
            """
            Initializes a new instance of the TicTacToe class.
            """
            self.board = [[' ' for _ in range(3)] for _ in range(3)]
            self.current_player = 'X'
            self.winner = None
            self.draw = False

        def is_valid_move(self, row, col):
            """
            Returns True if the specified row and column represent a valid move, False otherwise.

            Args:
                row (int): The row to check.
                col (int): The column to check.

            Returns:
                bool: True if the specified row and column represent a valid move, False otherwise.
            """
            return self.board[row][col] == ' '

        def check_winner(self):
            """
            Checks if the current player has won the game.

            Returns:
                str: The symbol of the player who has won the game, or None if the game has not been won yet.
            """
            # Check rows
            for row in range(3):
                if self.board[row][0] == self.board[row][1] == self.board[row][2] != ' ':
                    return self.board[row][0]

            # Check columns
            for col in range(3):
                if self.board[0][col] == self.board[1][col] == self.board[2][col] != ' ':
                    return self.board[0][col]

            # Check diagonals
            if self.board[0][0] == self.board[1][1] == self.board[2][2] != ' ':
                return self.board[0][0]
            if self.board[0][2] == self.board[1][1] == self.board[2][0] != ' ':
                return self.board[0][2]

            # No winner yet
            return None

        def check_draw(self):
            """
            Checks if the game has ended in a draw.

            Returns:
                bool: True if the game has ended in a draw, False otherwise.
            """
            for row in range(3):
                for col in range(3):
                    if self.board[row][col] == ' ':
                        return False
            return True

        def make_move(self, row, col):
            """
            Places the current player's symbol on the board at the specified row and column.
            If the move results in a win, sets the winner attribute to the current player.
            If the move results in a draw, sets the draw attribute to True.
            Switches to the other player's turn.

            Args:
                row (int): The row to place the symbol in.
                col (int): The column to place the symbol in.
            """
            # Check if the move is valid
            if self.is_valid_move(row, col):
                # Place the current player's symbol on the board
                self.board[row][col] = self.current_player
                # Check if the current move results in a win
                if self.check_winner() is not None:
                    self.winner = self.current_player
                # Check if the game has ended in a draw
                elif self.check_draw():
                    self.draw = True
                # Switch to the other player's turn
                self.current_player = 'X' if self.current_player == 'O' else 'O'

    def is_valid_move(self, row, col):
            """
            Check if the given move is valid.

            Args:
                row (int): The row of the move.
                col (int): The column of the move.

            Returns:
                bool: True if the move is valid, False otherwise.
            """
            # Check if the move is within the board boundaries and the cell is empty
            if 0 <= row < 3 and 0 <= col < 3 and self.board[row][col] == ' ':
                return True
            return False

    def check_winner(self):
        """
        Check if there is a winner in the current state of the board.

        Returns:
            str: The symbol of the winner ('X' or 'O') if there is a winner, otherwise None.
        """
        # Check for a win in rows
        ...
    def check_winner(self):
            """
            Check if there is a winner in the current state of the board.

            Returns:
                str: The symbol of the winner ('X' or 'O') if there is one, otherwise None.
            """
            # Check for a win in rows
            for row in self.board:
                if row[0] == row[1] == row[2] and row[0] != ' ':
                    return row[0]

            # Check for a win in columns
            for col in range(3):
                if self.board[0][col] == self.board[1][col] == self.board[2][col] and self.board[0][col] != ' ':
                    return self.board[0][col]

            # Check for a win in diagonals
            if self.board[0][0] == self.board[1][1] == self.board[2][2] and self.board[0][0] != ' ':
                return self.board[0][0]
            if self.board[0][2] == self.board[1][1] == self.board[2][0] and self.board[0][2] != ' ':
                return self.board[0][2]

            return None

    def check_draw(self):
            """
            Check if the game has ended in a draw.

            Returns:
                bool: True if all cells are filled, indicating a draw. False otherwise.
            """
            for row in self.board:
                if ' ' in row:
                    return False
            return True
